🛡️ Sentry: Fix panic risks in disambiguation and conversion
- Fixed panic in `src/semantic/disambiguation.rs` when sorting `MorphAnalysis` with `NaN` confidence by replacing `unwrap()` with `unwrap_or(Ordering::Equal)`.
- Replaced potentially unsafe `exprs.pop().unwrap()` in `src/semantic/conversion.rs` with `if let Some(...)` pattern.
- Added inline regression tests in `src/semantic/disambiguation.rs` and `src/semantic/conversion.rs` to verify safe handling of `NaN` confidence scores and empty expression propagation.
- Updated `.jules/sentry.md` with new learnings about recurring panic patterns.
